fix: remove old + rename
