+++ examples/stage_recorder.py
@@ -3,9 +3,10 @@
 import useq
 from pymmcore_plus import CMMCorePlus
 
+from pymmcore_widgets._stage_recorder import StageRecorder
+
 # from pymmcore_widgets import PropertyBrowser, MDAWidget, GroupPresetTableWidget
-from pymmcore_widgets._stage_recorder_no_images import StageTracker
-from pymmcore_widgets._stage_recorder_with_images import StageRecorder
+from pymmcore_widgets._stage_tracker import StageTracker
 
 mmc = CMMCorePlus().instance()
 cfg = r"c:\Users\NIC\Desktop\mm\Ti2.cfg"
