+++ src/pymmcore_widgets/_stage_recorder.py
