+++ src/pymmcore_widgets/_stage_tracker.py
